refactor(web): break ThemeTileGrid shell/hook import cycle via constants module
