+++ apps/web/src/components/shared/theme/ThemeTileGrid/index.ts
@@ -1,3 +1,3 @@
 export { default as ThemeTileGrid } from './ThemeTileGrid';
-export { THEME_TILES } from './ThemeTileGrid';
+export { THEME_TILES } from './ThemeTileGrid.constants';
 export * from './ThemeTileGrid.types';
